requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1
django-cors-headers==3.7.0
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2==2.9
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.3.0
